ui/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
umap-learn>=0.5.3
numpy>=1.24.0
